+++ Chord Dictator/MainWindow.xaml.cs
@@ -314,7 +314,7 @@ namespace Chord_Dictator
                     if (ofd.FileName.Substring(ofd.FileName.Length - 4, 4) == ".txt")
                     {
                         dfp = dictionaryPath + GetFileName(ofd.FileName);
-                        File.Copy(ofd.FileName, dfp);
+                        if (!File.Exists(dfp)) File.Copy(ofd.FileName, dfp);
                         FileOk(true);
 
                         MessageBox.Show("Dictionary file changed to " + dfp);
@@ -360,6 +360,13 @@ namespace Chord_Dictator
                 ofd.ShowDialog();
                 if (ofd.FileName.Length > 3)
                 {
+                    dfp = dictionaryPath + GetFileName(ofd.FileName);
+                    if (!File.Exists(dfp)) File.Copy(ofd.FileName, dfp);
+                    FileOk(true);
+
+                    MessageBox.Show("Dictionary file changed to " + dfp);
+                    WriteToLog("Dictionary file changed to " + dfp, "btnChangeInit_Click");
+                    btnGoToAdd.ToolTip = "All added items will be saved to " + dfp;
                     if (ofd.FileName.Substring(ofd.FileName.Length - 4, 4) == ".txt")
                     {
                         dfp = ofd.FileName;
